Updated app.js
I made it split up the title in parts. so it was easier to read everything. This was because it did have a country it just didn't know what to do with it.

Then I also made it possible to have country codes. So it transalates ceratin country names such as "United States" into "US" and if the name of the country isnt in the list it will fallback on using just the full country name